tools/python/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
boto3==1.34.0
deepseek-ai==0.1.0
openai==1.3.0 